download testcases separated from other handler tests.
